tools/google-docs-sync/requirements.txt
google-auth==2.32.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.134.0
googleapis-common-protos==1.63.2